requirements.dev.txt
django>=5.1.3,<6.0.0
djangorestframework>=3.15.2,<4.0.0
psycopg2-binary>=2.9.9,<3.0
drf-spectacular>=0.27.2,<0.28.0
ruff==0.5.7
django-cors-headers>=3.10,<4.0
djangorestframework_simplejwt>=5.2.2,<6.0.0